Compatibility with RegionScore plugin, store settings, iitc standard
Made this plugin to be runnable along side the original RegionScore
plugin.
The selected S2 Level is now stored and reused when Intel is reloaded.
Changed the button to be in the standard IITC Toolbox on the right, now
named "Change S2 Level"
Ran the code through jsbeautifier.org with 2-space-indent, and brace on
new line settings for easier future diffing